+++ src/api/sdk/index.ts
@@ -1,4 +1,3 @@
-// Copyright (c) Siemens AG, 2019
 export * from "./agent/agent-management";
 export * from "./agent/agent-models";
 export * from "./asset/asset-management";
@@ -16,6 +15,8 @@ export * from "./identity/identity";
 export * from "./identity/identity-models";
 export * from "./iot/iot-timeseries";
 export * from "./iot/iot-timeseries-models";
+export * from "./iotaggregate-v4/iot-timeseries-aggregate-models-v4";
+export * from "./iotaggregate-v4/iot-timeseries-aggregate-v4";
 export * from "./iotaggregate/iot-timeseries-aggregate";
 export * from "./iotaggregate/iot-timeseries-aggregate-models";
 export * from "./iotbulk/iot-timeseries-bulk";
@@ -24,6 +25,8 @@ export * from "./iotfile/iot-file";
 export * from "./iotfile/iot-file-models";
 export * from "./mcapi/mcapi";
 export * from "./mcapi/mcapi-models";
+export * from "./model/model-management";
+export * from "./model/model-models";
 export * from "./signal-validation/signal-validation";
 export * from "./signal-validation/signal-validation-models";
 export * from "./tenant/tenant-management";
@@ -32,5 +35,3 @@ export * from "./trend/trend-prediction";
 export * from "./trend/trend-prediction-models";
 export * from "./utc/utc";
 export * from "./utc/utc-models";
-export * from "./model/model-management";
-export * from "./model/model-models";